🚚 move : move hooks and store into store folder
