+++ src/app/lib/store/store.ts
@@ -1,7 +1,7 @@
 import { configureStore } from "@reduxjs/toolkit";
-import { authApi } from "./auth/authApi";
-import authReducer from "./auth/authSlice";
-import resetPasswordReducer from "./reset-password/resetPasswordSlice";
+import { authApi } from "../auth/authApi";
+import authReducer from "../auth/authSlice";
+import resetPasswordReducer from "../reset-password/resetPasswordSlice";
 
 export const store = configureStore({
     reducer: {
